poetry analyzer/pyproject.toml
[project]
name = "poetry-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents==0.2.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
